[REFACTOR] #87 - re-packaging
