--- app/src/main/java/co/kr/bemyplan/data/repository/list/NewListRepository.kt
@@ -1,7 +1,0 @@
-package co.kr.bemyplan.data.repository.list
-
-import co.kr.bemyplan.data.entity.list.ResponseNewList
-
-interface NewListRepository {
-    suspend fun getNewList(page: Int, pageSize: Int): ResponseNewList
-}
