+++ app/src/main/java/co/kr/bemyplan/data/repository/list/preview/PreviewInfoRepository.kt
@@ -1,4 +1,4 @@
-package co.kr.bemyplan.data.repository.list
+package co.kr.bemyplan.data.repository.list.preview
 
 import co.kr.bemyplan.data.entity.purchase.before.ResponsePreviewInfo
