+++ app/src/main/java/co/kr/bemyplan/data/repository/list/userpost/UserPostListRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package co.kr.bemyplan.data.repository.list
+package co.kr.bemyplan.data.repository.list.userpost
 
 import co.kr.bemyplan.data.api.ApiService
 import co.kr.bemyplan.data.entity.list.ResponseUserPostList
