+++ app/src/main/java/co/kr/bemyplan/ui/list/viewmodel/ListViewModel.kt
@@ -6,10 +6,10 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import co.kr.bemyplan.data.entity.list.ContentModel
-import co.kr.bemyplan.data.repository.list.LocationListRepositoryImpl
-import co.kr.bemyplan.data.repository.list.NewListRepositoryImpl
-import co.kr.bemyplan.data.repository.list.SuggestListRepositoryImpl
-import co.kr.bemyplan.data.repository.list.UserPostListRepositoryImpl
+import co.kr.bemyplan.data.repository.list.location.LocationListRepositoryImpl
+import co.kr.bemyplan.data.repository.list.newest.NewListRepositoryImpl
+import co.kr.bemyplan.data.repository.list.suggest.SuggestListRepositoryImpl
+import co.kr.bemyplan.data.repository.list.userpost.UserPostListRepositoryImpl
 import co.kr.bemyplan.data.repository.main.scrap.ScrapListRepositoryImpl
 import kotlinx.coroutines.launch
