+++ app/src/main/java/co/kr/bemyplan/ui/purchase/before/viewmodel/BeforeChargingViewModel.kt
@@ -6,8 +6,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import co.kr.bemyplan.data.entity.purchase.before.ContentModel
 import co.kr.bemyplan.data.entity.purchase.before.PreviewInfoModel
-import co.kr.bemyplan.data.repository.list.PreviewInfoRepositoryImpl
-import co.kr.bemyplan.data.repository.list.PreviewListRepositoryImpl
+import co.kr.bemyplan.data.repository.list.preview.PreviewInfoRepositoryImpl
+import co.kr.bemyplan.data.repository.list.preview.PreviewListRepositoryImpl
 import kotlinx.coroutines.launch
 
 class BeforeChargingViewModel: ViewModel() {
